rename to propTypes
