+++ src/pages/signin/SignInPageLayout/BackgroundImage/index.js
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types';
 import MobileBackgroundImage from '../../../../../assets/images/home-background--mobile.svg';
 import DesktopBackgroundImage from '../../../../../assets/images/home-background--desktop.svg';
 import styles from '../../../../styles/styles';
-import defaultPropTypes from './defaultPropTypes';
+import defaultPropTypes from './propTypes';
 
 const defaultProps = {
     isSmallScreen: false,
